Refactor code and improve app update process
This commit refactors several areas for code clarity and improves the process for updating the app. In particular, the update process now features separate, reusable functions for fetching and reloading an update, as well as better error handling. Unnecessary imports have also been removed and there are some improvements to the UI structure too.