Add CarouselItemUsers component and enhance UI
The CarouselItemUsers component was introduced to display user info in a carousel layout. Several files related to routes were removed to clean up the codebase. UI changes were made to the CongressoPage and InfoCongressCarousel components for improved presentation and readability. Dependencies in package.json were also updated for compatibility and functionality improvement.